services/elections/requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
psycopg2-binary==2.7.7
Flask-Testing==0.7.1
gunicorn==19.9.0